feat: added country exposure filter
Note that we filter country_exposure preemptively (in preprocess.R) instead of waiting until dim-reduce.R to lessen overhead for kcount. Do note that we defer dim-reduce factor filters to pre_reduce because we want the data to be more manageable as kmer matrices with metadata instead of separate FASTA and metadata which requires maintenance of drop indexes.